bsp/MicroLink/microlink_app/hpm_sdk_localized_for_hpm5301evklite/scripts/requirements.txt
PyYAML
jinja2
